# openapi/dindin.yaml
...
openapi: '3.1.0'
info:
  title: DinDin API
  version: '1.0.0'
  description: |
    API do DinDin — carteiras, posições, geladeira, proventos, patrimônio,
    carteira sugerida, simulação e cobrança.
servers:
  - url: https://dindin-4e720.web.app
    description: Produção
  - url: http://localhost:5001
    description: Emuladores
security:
  - firebaseIdToken: []
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ID token do Firebase Auth, enviado em toda requisição.
  responses:
    Unauthorized:
      description: Sem ID token válido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NoContent:
      description: Sem corpo.
  schemas:
...
    MeResponse:
      type: object
      required: [uid, admin, subscription, entitlements]
      properties:
        uid: { type: string }
        admin: { type: boolean }
        subscription: { $ref: '#/components/schemas/PublicSubscription' }
        entitlements:
          type: array
          items: { $ref: '#/components/schemas/Entitlement' }
...
    SetupRequest:
      type: object
      description: 'Corpo opcional: pedido explícito pelo fallback.'
      properties:
        resource: { $ref: '#/components/schemas/DefaultResource' }
...
    RegisterDeviceTokenRequest:
      type: object
      description: |
        Token de notificação do aparelho (issue #408). O app o registra a cada
        abertura; o backend atualiza o existente em vez de duplicar.
      required: [token, platform]
      properties:
        token: { type: string }
        platform: { $ref: '#/components/schemas/DevicePlatform' }
    SetupResponse:
      type: object
      required: [walletCreated, fridgeCreated]
      properties:
        walletCreated: { type: boolean }
        fridgeCreated: { type: boolean }
...
    CheckoutSessionResponse:
      type: object
      required: [url]
      properties:
        url: { type: string }
...
paths:
...
  /api/me:
    get:
      summary: Usuário autenticado, assinatura e entitlements
      operationId: getMe
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me/setup:
    post:
      summary: Cria os recursos padrão do usuário
      operationId: setupDefaults
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '200':
          description: O que foi criado nesta chamada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me/notification-tokens:
    post:
      summary: Registra o token de notificação do aparelho
      operationId: registerNotificationToken
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterDeviceTokenRequest' }
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/me/notification-tokens/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema: { type: string }
    delete:
      summary: Remove o token de notificação
      description: |
        É como o usuário desliga as notificações push dentro do app, sem
        depender das configurações do sistema.
      operationId: removeNotificationToken
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/billing/webhook:
    post:
      summary: Webhook da Stripe
      operationId: stripeWebhook
      security: []
      responses:
        '200': { description: Evento recebido }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/billing/checkout-session:
    post:
      summary: Abre a sessão de checkout da assinatura
      operationId: createCheckoutSession
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutSessionRequest' }
      responses:
        '200':
          description: URL do checkout
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/billing/portal-session:
    post:
      summary: Abre o portal de gestão da assinatura
      operationId: createPortalSession
      responses:
        '200':
          description: URL do portal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/billing/store/purchase:
    post:
      summary: Valida a compra na loja e concede a assinatura (também restaura)
      operationId: registerStorePurchase
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StorePurchaseRequest' }
      responses:
        '200':
          description: Assinatura depois da validação do recibo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicSubscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: 'Assinatura já ativa (`ALREADY_SUBSCRIBED`) ou recibo de outra conta (`RECEIPT_ALREADY_USED`)'
        '503':
          description: 'Loja não configurada (`STORE_NOT_CONFIGURED`)'
  /api/billing/store/apple-notifications:
    post:
      summary: Notificação de servidor da App Store
      operationId: appleStoreNotification
      security: []
      responses:
        '200': { description: Notificação recebida }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/billing/store/google-notifications:
    post:
      summary: Notificação de servidor do Google Play (RTDN)
      operationId: googleStoreNotification
      security: []
      responses:
        '200': { description: Notificação recebida }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/assets:
    get:
      summary: Catálogo de ativos ativos
      operationId: listAssets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Asset' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/admin/assets:
    get:
      summary: Catálogo completo (admin)
      operationId: adminListAssets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Asset' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Cadastra um ativo (admin)
      operationId: adminCreateAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAssetRequest' }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/admin/assets/{ticker}:
    parameters:
      - name: ticker
        in: path
        required: true
        schema: { type: string }
    put:
      summary: Atualiza um ativo (admin)
      operationId: adminUpdateAsset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAssetRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/users:
    get:
      summary: Usuários e assinaturas (admin)
      operationId: adminListUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/admin/users/{uid}/subscription:
    parameters:
      - name: uid
        in: path
        required: true
        schema: { type: string }
    put:
      summary: Concede assinatura manual (admin)
      operationId: adminGrantSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GrantSubscriptionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '400': { $ref: '#/components/responses/BadRequest' }
    delete:
      summary: Revoga a concessão manual (admin)
      operationId: adminRevokeSubscription
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminUser' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/quotes/dividend-history:
    get:
      summary: Histórico de proventos de vários tickers
      operationId: getDividendHistoryBatch
      parameters:
        - name: tickers
          in: query
          required: true
          schema: { type: string }
          description: Tickers separados por vírgula.
        - name: months
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                { $ref: '#/components/schemas/DividendHistoryBatchResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/quotes/{ticker}/dividend-history:
    get:
      summary: Histórico de proventos de um ticker
      operationId: getDividendHistory
      parameters:
        - name: ticker
          in: path
          required: true
          schema: { type: string }
        - name: months
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DividendHistoryResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/wallets:
    get:
      summary: Carteiras do usuário
      operationId: listWallets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Wallet' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Cria uma carteira
      operationId: createWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWalletRequest' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/wallets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Uma carteira
      operationId: getWallet
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Atualiza a carteira
      operationId: updateWallet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateWalletRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wallet' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Exclui a carteira
      operationId: deleteWallet
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/monthly-income:
    get:
      summary: Renda mensal projetada consolidada
      operationId: getConsolidatedMonthlyIncome
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthlyIncomeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/dashboard/summary:
    get:
      summary: Resumo do patrimônio e da composição
      operationId: getDashboardSummary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardSummaryResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/wallets/{walletId}/dividend-yield:
    parameters:
      - name: walletId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Dividend yield da carteira
      operationId: getDividendYield
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DividendYieldResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/wallets/{walletId}/monthly-income:
    parameters:
      - name: walletId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Renda mensal projetada da carteira
      operationId: getMonthlyIncome
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthlyIncomeResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/wallets/{walletId}/positions:
    parameters:
      - name: walletId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Posições da carteira
      operationId: listPositions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      summary: Cria uma posição
      operationId: createPosition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePositionRequest' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/wallets/{walletId}/positions/{id}:
    parameters:
      - name: walletId
        in: path
        required: true
        schema: { type: string }
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Uma posição
      operationId: getPosition
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Atualiza a posição
      operationId: updatePosition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePositionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Position' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Exclui a posição
      operationId: deletePosition
      responses:
        '204': { $ref: '#/components/responses/NoContent' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/dividends/projection:
    get:
      summary: Projeção de proventos
      operationId: getDividendProjection
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MonthlyIncomeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...